experiments/genai_devrel_weekly_challenge/week10/pyproject.toml
[project]
name = "week10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "google-auth-oauthlib>=1.2.1",
    "llm-guard>=0.3.15",
    "pip>=25.0.1",
    "pydantic-settings[yaml]>=2.8.1",
    "vertexai>=1.71.1",
]
